FUDGE-6.5.0 release. Changes include 1) fixing multiple translation and processing issues for ENDF-VIII.1 candidate evaluations, 2) expanding testing tools used to help check for quality issues in ENDF-VIII.1 candidates,  3) cleaning up some unused classes, especially in PoPs,  4) refining the PoPs 'intid' capability to match the implementation in GIDIplus, 5) fixing several issues with ACE production brownies/LANL/toACE, 6) expanding documentation, assorted bug fixes and minor improvements.

+++ xData/xDataArray.py
@@ -119,7 +119,7 @@ class ArrayBase( baseModule.XDataCoreMembers ) :
         """
 
         if self.dimension == 2:
-            assert secondIndex is not None and isinstance(secondIndex, numpy.int), 'The method constructVector requires the first argument to be an integer.'
+            assert secondIndex is not None and isinstance(secondIndex, int), 'The method constructVector requires the first argument to be an integer.'
 
             if self.shape[1] <= secondIndex:
                 return vectorModule.Vector()
@@ -144,7 +144,7 @@ class ArrayBase( baseModule.XDataCoreMembers ) :
         if self.dimension != 3:
             raise TypeError(f'The method constructMatrix is only defined for 3D arrays')
 
-        assert thirdIndex is not None and isinstance(thirdIndex, numpy.int), 'The method constructMatrix requires the first argument to be an integer.'
+        assert thirdIndex is not None and isinstance(thirdIndex, int), 'The method constructMatrix requires the first argument to be an integer.'
         if self.shape[2] <= thirdIndex:
             return matrixModule.Matrix()
